ams2/ams2-ai-livery-designer/python-backend/services/image_processor.py
"""
Image Processing Service
Handles photo preprocessing, background removal, and image transformations
"""

import numpy as np
from PIL import Image
import cv2
from typing import Optional, Tuple


class ImageProcessor:
    """
    Service for processing reference photos before AI generation
    
    Features (Week 3-4):
    - Background removal using Segment Anything Model (SAM)
    - Image normalization and color correction
    - Multi-view photo alignment
    - Perspective correction
    - Photo quality assessment
    """
    
    def __init__(self):
        """Initialize image processor"""
        self.sam_model = None  # Will load SAM in Week 3
        
    def load_sam_model(self):
        """Load Segment Anything Model for background removal"""
        # TODO: Implement in Week 3
        # from segment_anything import sam_model_registry, SamPredictor
        # self.sam_model = sam_model_registry["vit_h"](checkpoint="sam_vit_h.pth")
        pass
    
    def remove_background(self, image: np.ndarray) -> np.ndarray:
        """
        Remove background from car photo using SAM
        
        Args:
            image: Input image as numpy array (RGB)
            
        Returns:
            Image with background removed (RGBA with transparency)
        """
        # TODO: Implement SAM background removal in Week 3
        # For now, return original image
        return image
    
    def normalize_image(self, image: np.ndarray) -> np.ndarray:
        """
        Normalize image (color correction, exposure adjustment)
        
        Args:
            image: Input image as numpy array
            
        Returns:
            Normalized image
        """
        # Basic normalization
        image = cv2.normalize(image, None, 0, 255, cv2.NORM_MINMAX)
        return image
    
    def detect_car_angle(self, image: np.ndarray) -> str:
        """
        Detect viewing angle of car photo (front/side/rear/3-4)
        
        Args:
            image: Input image as numpy array
            
        Returns:
            Detected angle: "front", "side", "rear", "3/4", or "unknown"
        """
        # TODO: Implement ML-based angle detection in Week 4
        return "unknown"
    
    def assess_quality(self, image: np.ndarray) -> dict:
        """
        Assess photo quality for livery generation
        
        Args:
            image: Input image as numpy array
            
        Returns:
            Quality metrics dict with scores (0-1)
        """
        height, width = image.shape[:2]
        
        # Basic quality checks
        quality = {
            "resolution": 1.0 if width >= 1920 else width / 1920,
            "sharpness": 0.8,  # TODO: Implement blur detection
            "lighting": 0.9,   # TODO: Implement lighting analysis
            "overall": 0.85    # Average of all metrics
        }
        
        return quality
    
    def preprocess_for_ai(
        self, 
        image: np.ndarray,
        target_size: Tuple[int, int] = (1024, 1024)
    ) -> np.ndarray:
        """
        Preprocess image for AI generation (resize, normalize)
        
        Args:
            image: Input image
            target_size: Target dimensions for AI model
            
        Returns:
            Preprocessed image ready for SDXL/ControlNet
        """
        # Resize to target size
        image = cv2.resize(image, target_size, interpolation=cv2.INTER_LANCZOS4)
        
        # Normalize to [0, 1] range
        image = image.astype(np.float32) / 255.0
        
        return image
