ams2/ams2-ai-livery-designer/python-backend/requirements.txt
# Python Backend Requirements - AMS2 AI Livery Designer
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PyTorch with CUDA (matching POC-08)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# AI/ML - Week 2 Installation
# Stable Diffusion XL
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# ControlNet & IPAdapter - Week 2
# controlnet-aux>=0.0.7
# insightface>=0.7.3  # For IPAdapter

# Segment Anything Model (SAM) - Week 3
# segment-anything>=1.0
# git+https://github.com/facebookresearch/segment-anything.git

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# DDS Texture Support - Week 6
# pydds  # DDS file format handling

# Utilities
requests>=2.31.0
huggingface-hub>=0.19.0
psutil>=5.9.0

# Development
python-dotenv>=1.0.0
